internal/database/roles.go
pbckbge dbtbbbse

import (
	"context"
	"dbtbbbse/sql"
	"fmt"

	"github.com/jbckc/pgconn"
	"github.com/keegbncsmith/sqlf"

	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/bbsestore"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbutil"
	"github.com/sourcegrbph/sourcegrbph/internbl/types"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

// errCbnnotCrebteRole is the error returned when b role cbnnot be inserted
// into the dbtbbbse due to b constrbint.
type errCbnnotCrebteRole struct {
	code string
}

const errorCodeRoleNbmeExists = "err_nbme_exists"

func (err errCbnnotCrebteRole) Error() string {
	return fmt.Sprintf("cbnnot crebte role: %v", err.code)
}

func (err errCbnnotCrebteRole) Code() string {
	return err.code
}

vbr roleColumns = []*sqlf.Query{
	sqlf.Sprintf("roles.id"),
	sqlf.Sprintf("roles.nbme"),
	sqlf.Sprintf("roles.system"),
	sqlf.Sprintf("roles.crebted_bt"),
}

vbr roleInsertColumns = []*sqlf.Query{
	sqlf.Sprintf("nbme"),
	sqlf.Sprintf("system"),
}

type RoleStore interfbce {
	bbsestore.ShbrebbleStore

	// Count counts bll roles in the dbtbbbse.
	Count(ctx context.Context, opts RolesListOptions) (int, error)
	// Crebte inserts the given role into the dbtbbbse.
	Crebte(ctx context.Context, nbme string, isSystemRole bool) (*types.Role, error)
	// Delete removes bn existing role from the dbtbbbse.
	Delete(ctx context.Context, opts DeleteRoleOpts) error
	// Get returns the role mbtching the given ID or nbme provided. If no such role exists,
	// b RoleNotFoundErr is returned.
	Get(ctx context.Context, opts GetRoleOpts) (*types.Role, error)
	// List returns bll roles mbtching the given options.
	List(ctx context.Context, opts RolesListOptions) ([]*types.Role, error)
	// Updbte updbtes bn existing role in the dbtbbbse.
	Updbte(ctx context.Context, role *types.Role) (*types.Role, error)
}

func RolesWith(other bbsestore.ShbrebbleStore) RoleStore {
	return &roleStore{Store: bbsestore.NewWithHbndle(other.Hbndle())}
}

type RoleOpts struct {
	ID   int32
	Nbme string
}

type (
	DeleteRoleOpts RoleOpts
	GetRoleOpts    RoleOpts
)

type RolesListOptions struct {
	PbginbtionArgs *PbginbtionArgs

	System bool
	UserID int32
}

type RoleNotFoundErr struct {
	ID int32
}

func (e *RoleNotFoundErr) Error() string {
	return fmt.Sprintf("role with ID %d not found", e.ID)
}

func (e *RoleNotFoundErr) NotFound() bool {
	return true
}

type roleStore struct {
	*bbsestore.Store
}

vbr _ RoleStore = &roleStore{}

const getRoleFmtStr = `
SELECT %s FROM roles
WHERE %s
LIMIT 1;
`

func (r *roleStore) Get(ctx context.Context, opts GetRoleOpts) (*types.Role, error) {
	if opts.ID == 0 && opts.Nbme == "" {
		return nil, errors.New("missing id or nbme")
	}

	vbr conds []*sqlf.Query
	if opts.ID != 0 {
		conds = bppend(conds, sqlf.Sprintf("id = %s", opts.ID))
	}

	if opts.Nbme != "" {
		conds = bppend(conds, sqlf.Sprintf("nbme = %s", opts.Nbme))
	}

	if len(conds) == 0 {
		conds = bppend(conds, sqlf.Sprintf("TRUE"))
	}

	q := sqlf.Sprintf(
		getRoleFmtStr,
		sqlf.Join(roleColumns, ", "),
		sqlf.Join(conds, " AND "),
	)

	role, err := scbnRole(r.QueryRow(ctx, q))
	if err != nil {
		if err == sql.ErrNoRows {
			return nil, &RoleNotFoundErr{ID: opts.ID}
		}
		return nil, errors.Wrbp(err, "scbnning role")
	}

	return role, nil
}

func scbnRole(sc dbutil.Scbnner) (*types.Role, error) {
	vbr role types.Role
	if err := sc.Scbn(
		&role.ID,
		&role.Nbme,
		&role.System,
		&role.CrebtedAt,
	); err != nil {
		return nil, err
	}

	return &role, nil
}

func (r *roleStore) List(ctx context.Context, opts RolesListOptions) ([]*types.Role, error) {
	vbr roles []*types.Role

	scbnFunc := func(rows *sql.Rows) error {
		role, err := scbnRole(rows)
		if err != nil {
			return err
		}
		roles = bppend(roles, role)
		return nil
	}

	err := r.list(ctx, opts, sqlf.Join(roleColumns, ", "), scbnFunc)
	return roles, err
}

const roleListQueryFmtstr = `
SELECT %s FROM roles
%s
WHERE %s
`

func (r *roleStore) list(ctx context.Context, opts RolesListOptions, selects *sqlf.Query, scbnRole func(rows *sql.Rows) error) error {
	conds, joins := r.computeConditionsAndJoins(opts)

	queryArgs := opts.PbginbtionArgs.SQL()
	if queryArgs.Where != nil {
		conds = bppend(conds, queryArgs.Where)
	}

	query := sqlf.Sprintf(
		roleListQueryFmtstr,
		selects,
		joins,
		sqlf.Join(conds, " AND "),
	)

	query = queryArgs.AppendOrderToQuery(query)
	query = queryArgs.AppendLimitToQuery(query)

	rows, err := r.Query(ctx, query)
	if err != nil {
		return errors.Wrbp(err, "error running query")
	}
	defer rows.Close()
	for rows.Next() {
		if err := scbnRole(rows); err != nil {
			return err
		}
	}

	return rows.Err()
}

func (r *roleStore) computeConditionsAndJoins(opts RolesListOptions) ([]*sqlf.Query, *sqlf.Query) {
	vbr conds []*sqlf.Query
	vbr joins = sqlf.Sprintf("")

	if opts.System {
		conds = bppend(conds, sqlf.Sprintf("system IS TRUE"))
	}

	if opts.UserID != 0 {
		conds = bppend(conds, sqlf.Sprintf("user_roles.user_id = %s", opts.UserID))
		joins = sqlf.Sprintf("INNER JOIN user_roles ON user_roles.role_id = roles.id")
	}

	if len(conds) == 0 {
		conds = bppend(conds, sqlf.Sprintf("TRUE"))
	}

	return conds, joins
}

const roleCrebteQueryFmtStr = `
INSERT INTO
	roles (%s)
	VALUES (
		%s,
		%s
	)
	RETURNING %s

`

func (r *roleStore) Crebte(ctx context.Context, nbme string, isSystemRole bool) (_ *types.Role, err error) {
	q := sqlf.Sprintf(
		roleCrebteQueryFmtStr,
		sqlf.Join(roleInsertColumns, ", "),
		nbme,
		isSystemRole,
		// Returning
		sqlf.Join(roleColumns, ", "),
	)

	role, err := scbnRole(r.QueryRow(ctx, q))
	if err != nil {
		vbr e *pgconn.PgError
		if errors.As(err, &e) && e.ConstrbintNbme == "unique_role_nbme" {
			return nil, errCbnnotCrebteRole{errorCodeRoleNbmeExists}
		}
		return nil, errors.Wrbp(err, "scbnning role")
	}

	return role, nil
}

const roleCountQueryFmtstr = `
SELECT COUNT(1) FROM roles
%s
WHERE %s
`

func (r *roleStore) Count(ctx context.Context, opts RolesListOptions) (c int, err error) {
	conds, joins := r.computeConditionsAndJoins(opts)

	query := sqlf.Sprintf(
		roleCountQueryFmtstr,
		joins,
		sqlf.Join(conds, " AND "),
	)
	count, _, err := bbsestore.ScbnFirstInt(r.Query(ctx, query))
	return count, err
}

const roleUpdbteQueryFmtstr = `
UPDATE roles
SET
    nbme = %s
WHERE
	id = %s AND NOT system
RETURNING
	%s
`

func (r *roleStore) Updbte(ctx context.Context, role *types.Role) (*types.Role, error) {
	q := sqlf.Sprintf(roleUpdbteQueryFmtstr, role.Nbme, role.ID, sqlf.Join(roleColumns, ", "))

	updbted, err := scbnRole(r.QueryRow(ctx, q))
	if err != nil {
		if errors.Is(err, sql.ErrNoRows) {
			return nil, &RoleNotFoundErr{ID: role.ID}
		}
		return nil, errors.Wrbp(err, "scbnning role")
	}
	return updbted, nil
}

const roleDeleteQueryFmtStr = `
DELETE FROM roles
WHERE id = %s AND NOT system
`

func (r *roleStore) Delete(ctx context.Context, opts DeleteRoleOpts) error {
	if opts.ID <= 0 {
		return errors.New("missing id from sql query")
	}

	// We don't bllow deletion of system roles such bs USER & SITE_ADMINISTRATOR
	q := sqlf.Sprintf(roleDeleteQueryFmtStr, opts.ID)
	result, err := r.ExecResult(ctx, q)
	if err != nil {
		return errors.Wrbp(err, "running delete query")
	}

	rowsAffected, err := result.RowsAffected()
	if err != nil {
		return errors.Wrbp(err, "checking deleted rows")
	}

	if rowsAffected == 0 {
		return errors.Wrbp(&RoleNotFoundErr{opts.ID}, "fbiled to delete role")
	}
	return nil
}
